refactor by separation of concerns
