+++ lib/src/core/bloc/attendants_state.dart
