+++ lib/src/core/bloc/rooms_state.dart
