+++ lib/src/core/models/attendant_model.dart
