+++ lib/src/core/models/available_rooms_model.dart
