+++ lib/src/core/models/inivigilator_model.dart
