+++ lib/src/core/models/teaching_assistant_model.dart
