+++ lib/src/my_app.dart
@@ -1,0 +1,19 @@
+import 'package:flutter/material.dart';
+
+import 'screens/views/home.dart';
+
+class MyApp extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return MaterialApp(
+      title: 'CoE INVIGILTORS ATTENDANCE',
+      theme: ThemeData(
+        primaryColor: const Color(0xFFFFFFFF),
+        accentColor: const Color(0xFF9c27b0),
+        canvasColor: const Color(0xFFECEFF1),
+      ),
+      home: HomePage(),
+      debugShowCheckedModeBanner: false,
+    );
+  }
+}
