+++ lib/src/screens/components/footer.dart
